requirements.txt
altair==5.4.1
altgraph==0.17.4
arrow==1.3.0
attrs==24.2.0
beautifulsoup4==4.12.3
beeware==0.3.0
binaryornot==0.4.4
blinker==1.8.2
briefcase==0.3.19
build==1.2.2.post1
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cookiecutter==2.6.0
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
frozendict==2.4.6
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
html5lib==1.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
instaloader==4.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jpype1==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
moviepy==1.0.3
multitasking==0.0.11
mysql-connector-python==9.0.0
narwhals==1.10.0
numpy==1.24.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pdf2image==1.17.0
peewee==3.17.8
pefile==2023.2.7
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
proglog==0.1.10
protobuf==5.28.3
psutil==5.9.8
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.1.4
pyproject_hooks==1.2.0
pyTelegramBotAPI==4.23.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smmap==5.0.1
soupsieve==2.6
SQLAlchemy==2.0.37
streamlit==1.39.0
ta==0.11.0
tabula-py==2.9.3
telebot==0.0.5
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
tk==0.1.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.0.0
tornado==6.4.1
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
webencodings==0.5.1
Werkzeug==3.0.6
xlrd==2.0.1
XlsxWriter==3.2.0
yfinance==0.2.50
zipp==3.20.2
